refactor: split admin panel into components
